Remove the historical note about support being added in 3.2, value added at this point is questionable
Merge the note about supported versions and modes (Salt, not Traditional) in the general support statement.

Make the list of supported and unsupported features a bit more straightforward, and move below the details
about what we support for bootstrapping.

For IP reasons, add note about Canonical not supporting or endorsing SUSE Manager. We should have had
that since the very beginning:
https://ubuntu.com/legal/intellectual-property-policy